requirements.txt
# Core LangGraph and LLM Framework
langgraph>=0.2.6
langgraph-sdk>=0.1.61
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangChain Integrations
langchain-chroma>=0.1.0
langchain-google-genai>=2.0.0

# Web Framework and Server
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.20.0

# Data Processing and Validation
pydantic>=2.0.0
python-dotenv>=1.0.1
typing-extensions>=4.5.0

# Web Scraping and Browser Automation
zendriver>=1.0.0
aiohttp>=3.8.0
requests>=2.28.0

# Document Processing
pypdf2>=3.0.0
python-docx>=0.8.11

# Vector Database
chromadb>=0.4.0

# Development and Testing (Optional)
pytest>=7.0.0
mypy>=1.11.1
ruff>=0.6.1
